wsgi.py
"""WSGI entry point."""
import os
from dotenv import load_dotenv

# Load environment variables from .env file (for local development)
load_dotenv()

from app import create_app

app = create_app()

if __name__ == '__main__':
    port = int(os.environ.get('PORT', 8000))
    app.run(host='0.0.0.0', port=port, debug=os.environ.get('FLASK_ENV') == 'development')

